move testing to tox+pytest
tox replaces the manual environment setup and testting of installation in the Makefile.
Currently the actual tests are marked as "skip" for py.test, since they weren't being automatically executed anyway (and require an actual GLPI installation to test against), but in the future they could be made to use the mock module.

Tests can be enabled by setting the VCAP_SERVICES variable.